Created game folder WIP
